Master

+++ src/main/java/com/SDP/project/Repository/InventoryRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface InventoryRepository extends JpaRepository<Inventory, Long> {
+public interface InventoryRepository extends JpaRepository<Inventory, Integer> {
     Optional<Inventory> findByTaskId(int taskId);
 }
